Multiplayer implementation

+++ Assets/Scripts/Saving/Data/EffectsData/EffectData.cs
@@ -3,11 +3,11 @@
     public class EffectData
     {
         public string type;
-        public float duration;
+        public int duration;
         public float amplifier;
         public bool infinite;
 
-        public EffectData(float duration, float amplifier, bool infinite, string type)
+        public EffectData(int duration, float amplifier, bool infinite, string type)
         {
             this.infinite = infinite;
             this.duration = duration;
